requirements.txt
openai>=1.93.0
openai-agents>=0.1.0
pydantic>=2.0.0
fastapi>=0.104.0
requests>=2.31.0
typing-extensions>=4.8.0
python-dotenv>=1.0.0
websockets>=12.0
uvicorn>=0.24.0

# Development dependencies
pre-commit>=4.2.0
black>=24.4.0
ruff>=0.3.4
mypy>=1.9.0
